// src/generated/globalTypes.ts
/* tslint:disable */
/* eslint-disable */
// This file was automatically generated and should not be edited.

//==============================================================
// START Enums and Input Objects
//==============================================================

/**
 * The possible states of an issue.
 */
export enum IssueState {
  CLOSED = "CLOSED",
  OPEN = "OPEN",
}

/**
 * The possible states of a subscription.
 */
export enum SubscriptionState {
  IGNORED = "IGNORED",
  SUBSCRIBED = "SUBSCRIBED",
  UNSUBSCRIBED = "UNSUBSCRIBED",
}

//==============================================================
// END Enums and Input Objects
//==============================================================